base class for preprocesssing/analysis pipelines
