+++ subsystems/dpr/data_analysis_pipelines/base.py
@@ -1,4 +1,4 @@
 from subsystems.dpr.base_pipeline import BasePipeline
 
-class BasePreprocessingPipeline(BasePipeline):
+class BaseAnalysisPipeline(BasePipeline):
     pass
